+++ lib/sparse/general.h
@@ -61,7 +61,7 @@ double* vector_saxpy(int n, double *x, double *y, double beta); /* y = x+beta*y 
 double* vector_saxpy2(int n, double *x, double *y, double beta);/* x = x+beta*y */
 
 /* take m elements v[p[i]]],i=1,...,m and oput in u. u will be assigned if *u = NULL */
-void vector_float_take(int n, float *v, int m, int *p, float **u);
+void vector_float_take(int n, float *v, size_t m, int *p, float **u);
 
 /* give the position of the smallest, second smallest etc in vector v.
    results in p. If *p == NULL, p is assigned.
